FROMGIT: drm/i915/gem: Serialise debugfs i915_gem_objects with ctx->mutex
Since the debugfs may peek into the GEM contexts as the corresponding
client/fd is being closed, we may try and follow a dangling pointer.
However, the context closure itself is serialised with the ctx->mutex,
so if we hold that mutex as we inspect the state coupled in the context,
we know the pointers within the context are stable and will remain valid
as we inspect their tables.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: CQ Tang <cq.tang@intel.com>
Cc: Daniel Vetter <daniel.vetter@intel.com>
Cc: stable@vger.kernel.org
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200723172119.17649-3-chris@chris-wilson.co.uk
(cherry picked from commit 102f5aa491f262c818e607fc4fee08a724a76c69
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-gt-next)

BUG=b:169545458
TEST=Check kernel booted successfully without i915_gem_objects crash on dedede board

Signed-off-by: Mohanram Meenakshisundaram <mohanram.meenakshisundaram@intel.com>
Change-Id: Ibe839c66b3f57a554d512f2add7421b178f8a585
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2439599
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
Tested-by: Girish DS <girish.ds@intel.corp-partner.google.com>
Commit-Queue: Evan Green <evgreen@chromium.org>