CHROMIUM: iwl7000: mvm: properly cancel a session protection for P2P
We need to feed the configuration id to remove  session protection
properly.
Remember the conf_id when we add the session protection so that we
can give it back when we want to remove the session protection.
While at it, slightly improve the kernel doc for the conf_id
of the notification.

Change-Id: I01a98ecde62096d00d171cf34ad775bf80cb0277
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 0d2730992ace79c2dab55603fc1681be3905374e