CHROMIUM: config: Enable the TEO governor
The TEO governor is what all the cool kids are using.  Let's be cool
and enable it.

NOTE: Since "teo" has a rating of 19 and "menu" has a rating of 20, it
means that with both enabled we'll automagically switch anyone who was
using "menu" over to "teo" without the need for any kernel command
line parameters.

With this change, USB 2.0 reads on sc7180-trogdor were found to be 4%
faster.  Presumably other similar tasks will get a similar boost.

Other boards ran PLT (power load test) with this change and found
power impact (if any) OK.  However, boards should still keep an eye on
this change and confirm it had no bad impact.

BUG=b:147894727
TEST=PLT; check USB speeds

Change-Id: I3053c40f1fd98254eb17fde9b04e04e7fb96bbbc
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Disallow-Recycled-Builds: kukui-arc-r-cq
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2332918
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>