CHROMIUM: iwl7000: mvm: do more useful queue sync accounting
We're currently doing accounting on the queue sync with an
atomic variable that counts down the number of remaining
notifications that we still need.

As we've been hitting issues in this area, modify this to
track a bitmap of queues, not just the number of queues,
and print out the remaining bitmap in the warning.

Change-Id: I7c69ff999419368266279ec27dd618eb450908b3
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 1e2ee173065d0bab802db2d48e8c749c946a2e3f