FROMLIST: drm/msm/adreno: fix updating ring fence
We need to set it to the most recent completed fence, not the most
recent submitted.  Otherwise we have races where we think we can retire
submits that the GPU is not finished with, if the GPU doesn't manage to
overwrite the seqno before we look at it.

This can show up with hang recovery if one of the submits after the
crashing submit also hangs after it is replayed.

Fixes: f97decac5f4c ("drm/msm: Support multiple ringbuffers")
Signed-off-by: Rob Clark <robdclark@chromium.org>

(am from https://patchwork.freedesktop.org/patch/385284/)
Signed-off-by: Rob Clark <robdclark@chromium.org>

BUG=b:157648338
TEST=Trigger many GPU hangs with infinite loop shadertoy, verify we
     recover reliably

Change-Id: I2b8e9d8d446dd9143058f48cc43f29e2b812502e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2354653
Tested-by: Rob Clark <robdclark@chromium.org>
Auto-Submit: Rob Clark <robdclark@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Commit-Queue: Rob Clark <robdclark@chromium.org>