BACKPORT: opp: Reuse the enabled flag in !target_freq path
The OPP core needs to track if the resources of devices are
enabled/configured or not, as it disables the resources when target_freq
is set to 0.

Handle that with the new enabled flag and remove otherwise complex
conditional statements.

Tested-by: Rajendra Nayak <rnayak@codeaurora.org>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>

(cherry picked from commit 10b217365b9454bc5f31d2801148fdf04939f431)

Conflicts:
   drivers/opp/core.c
...conflict is because of difference between upstream commit
a4501bac0e55 ("opp: Enable resources again if they were disabled
earlier") and the commit that we got from stable merges, AKA commit
10e99c3003d5 ("opp: Enable resources again if they were disabled
earlier").

BUG=b:162365396
TEST=eMMC speeds are faster

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Icf78daad48cab9a0e99a64e7912b0fae280fcac7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2954426
Reviewed-by: Stephen Boyd <swboyd@chromium.org>