UPSTREAM: mei: add connect with vtag ioctl
This IOCTL is used to associate the current file descriptor
with a FW Client (given by UUID), and virtual tag (vtag).
The IOCTL opens a communication channel between a host client
and a FW client on a tagged channel. From this point on,
every reader  and write will communicate with the associated
FW client on the tagged channel. Upon close() the communication
is terminated.

The IOCTL argument is a struct with a union that contains
the input parameter and the output parameter for this IOCTL.

The input parameter is UUID of the FW Client, a vtag [0,255]
The output parameter is the properties of the FW client

Clients that do not support tagged connection
will respond with -EOPNOTSUPP

Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>
Link: https://lore.kernel.org/r/20200818115147.2567012-12-tomas.winkler@intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit aa207a05f95abc3530b7415232f0f73278336bd3)

BUG=b:155511178
TEST=None

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Id33713faed46ae29da48d36775454955f1e40e03
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2587324
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>