UPSTREAM: 53c700: improve non-coherent DMA handling
Switch the 53c700 driver to only use non-coherent descriptor memory if it
really has to because dma_alloc_coherent fails.  This doesn't matter for
any of the platforms it runs on currently, but that will change soon.

To help with this two new helpers to transfer ownership to and from the
device are added that abstract the syncing of the non-coherent memory.
The two current bidirectional cases are mapped to transfers to the
device, as that appears to what they are used for.  Note that for parisc,
which is the only architecture this driver needs to use non-coherent
memory on, the direction argument of dma_cache_sync is ignored, so this
will not change behavior in any way.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Tested-by: Thomas Bogendoerfer <tsbogend@alpha.franken.de>
(cherry picked from commit 91af2dd255ac9284ee2942a359fa8d2b49849e0b)

BUG=b:149115731
TEST=Run Camera app in lazor
TEST=Try https://appr.tc/?debug=loopback&vsc=vp8

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: Ic3920894b794822a0bd318decb304c3de5338369
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2773103
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>