UPSTREAM: virtiofs: add logic to free up a memory range
Add logic to free up a busy memory range. Freed memory range will be
returned to free pool. Add a worker which can be started to select
and free some busy memory ranges.

Process can also steal one of its busy dax ranges if free range is not
available. I will refer it to as direct reclaim.

If free range is not available and nothing can't be stolen from same
inode, caller waits on a waitq for free range to become available.

For reclaiming a range, as of now we need to hold following locks in
specified order.

	down_write(&fi->i_mmap_sem);
	down_write(&fi->dax->sem);

We look for a free range in following order.

A. Try to get a free range.
B. If not, try direct reclaim.
C. If not, wait for a memory range to become free

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>

(cherry picked from commit 9a752d18c85ae5da28e4a07d52adfd95eacb2495)

BUG=b:153701115
TEST=build

Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>
Change-Id: Ia209dda3404ad31badeb6ee303782d1366ce4474
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2539530
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>