UPSTREAM: dma-mapping: add a new dma_alloc_noncoherent API
Add a new API to allocate and free memory that is guaranteed to be
addressable by a device, but which potentially is not cache coherent
for DMA.

To transfer ownership to and from the device, the existing streaming
DMA API calls dma_sync_single_for_device and dma_sync_single_for_cpu
must be used.

For now the new calls are implemented on top of dma_alloc_attrs just
like the old-noncoherent API, but once all drivers are switched to
the new API it will be replaced with a better working implementation
that is available on all architectures.

Signed-off-by: Christoph Hellwig <hch@lst.de>
(cherry picked from commit 0d71675f87dc406d4c284729b8d36be050ad0d15)

BUG=b:149115731
TEST=Run Camera app in lazor
TEST=Try https://appr.tc/?debug=loopback&vsc=vp8

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: I7168bbf96ffc153b8aa8e1ddca5deca4a972fb13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2721959
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>