UPSTREAM: cpufreq: intel_pstate: Add ->offline and ->online callbacks
Add ->offline and ->online driver callbacks to prepare for taking a
CPU offline and to restore its working configuration when it goes
back online, respectively, to avoid invoking the ->init callback on
every CPU online which is quite a bit of unnecessary overhead.

Define ->offline and ->online so that they can be used in the
passive mode as well as in the active mode and because ->offline
will do the majority of ->stop_cpu work, the passive mode does
not need that callback any more, so drop it from there.

Also modify the active mode ->suspend and ->resume callbacks to
prevent them from interfering with the new ->offline and ->online
ones in case the latter are invoked withing the system-wide suspend
and resume code flow and make the passive mode use them too.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
(cherry picked from commit 4adcf2e5829f83cbab581ba376abb7cb6404eeb3)

BUG=b:135666007
TEST=Verify the sysfs interface with different values

Signed-off-by: Ed Baker <edward.baker@intel.com>
Signed-off-by: Sumeet R Pawnikar <sumeet.r.pawnikar@intel.com>
Change-Id: I7abe2b29403e5afb3b85db8e11b0d374c8a1acca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2444976
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>