UPSTREAM: cpufreq: intel_pstate: Free memory only when turning off
When intel_pstate switches the operation mode from "active" to
"passive" or the other way around, freeing its data structures
representing CPUs and allocating them again from scratch is not
necessary and wasteful.  Moreover, if these data structures are
preserved, the cached HWP Request MSR value from there may be
written to the MSR to start with to reinitialize it and help to
restore the EPP value set previously (it is set to 0xFF when CPUs
go offline to allow their SMT siblings to use the full range of
EPP values and that also happens when the driver gets unregistered).

Accordingly, modify the driver to only do a full cleanup on driver
object registration errors and when its status is changed to "off"
via sysfs and to write the cached HWP Request MSR value back to
the MSR on CPU init if the data structure representing the given
CPU is still there.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
(cherry picked from commit 55671ea3257ac596abd817c4031a996b2867e586)

BUG=b:135666007
TEST=Verify the sysfs interface with different values

Signed-off-by: Ed Baker <edward.baker@intel.com>
Signed-off-by: Sumeet R Pawnikar <sumeet.r.pawnikar@intel.com>
Change-Id: Idcb1e575ec77a05436e18d78a1048df23db407a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2444977
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>