BACKPORT: FROMGIT: mtd: spi-nor: Prefer asynchronous probe
On my system the spi_nor_probe() took ~6 ms at bootup.  That's not a
lot, but every little bit adds up to a slow bootup.  While we can get
this out of the boot path by making it a module, there are times where
it is convenient (or even required) for this to be builtin the kernel.
Let's set that we prefer async probe so that we don't block other
drivers from probing while we are probing.

This is a tiny little change that is almost guaranteed to be safe for
anything that is able to run as a module, which SPI_NOR is.
Specifically modules are already probed asynchronously.  Also: since
other things in the system may have enabled asynchronous probe the
system may already be doing other things during our probe.

There is a small possibility that some other driver that was a client
of SPI_NOR didn't handle -EPROBE_DEFER and was relying on probe
ordering and only worked when the SPI_NOR and the SPI bus were
builtin.  In that case the other driver has a bug that's waiting to
hit and the other driver should be fixed.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Link: https://lore.kernel.org/r/20200902160002.1.I658d1c0db9adfeb9a59bc55e96a19e192c959e55@changeid

(cherry picked from commit 03edda0e1edaa3c2e99239c66e3c14d749318fd6
 https://git.kernel.org/pub/scm/linux/kernel/git/mtd/linux.git spi-nor/next)

Conflicts:
   drivers/mtd/spi-nor/core.c
...the bit of code moved files upstream, but it's a 1-line change and
it's easy to find the new place.  Lots of reorganization happened
upstream and picking all that doesn't make sense.

BUG=b:166949233
TEST=Bootperf is slightly faster

Change-Id: I658d1c0db9adfeb9a59bc55e96a19e192c959e55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2439392
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Auto-Submit: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>