CHROMIUM: USB: misc: Add onboard_usb_hub driver
 [This patch was posted for upstream review and was well received,
  however the DT bindings were refused. The latest posted version
  is https://lore.kernel.org/patchwork/patch/1313001/.

  Since USB maintainers are generally supportive of the idea we intend to
  land the functionality eventually, but the final implementation will
  likely look different.

  Changes with respect to v4 posted on the list:
   - use late suspend to make sure info about wakeup enabled descendants
     is updated
   - renamed sysfs attribute 'power_off_in_suspend' to
     'always_powered_in_suspend'
   - fix log for regulator_disable() failure
   - removed 'realtek' compatible string which is not needed at this
     point
 ]

The main issue this driver addresses is that a USB hub needs to be
powered before it can be discovered. For discrete onboard hubs (an
example for such a hub is the Realtek RTS5411) this is often solved
by supplying the hub with an 'always-on' regulator, which is kind
of a hack. Some onboard hubs may require further initialization
steps, like changing the state of a GPIO or enabling a clock, which
requires even more hacks. This driver creates a platform device
representing the hub which performs the necessary initialization.
Currently it only supports switching on a single regulator, support
for multiple regulators or other actions can be added as needed.
Different initialization sequences can be supported based on the
compatible string.

Besides performing the initialization the driver can be configured
to power the hub off during system suspend. This can help to extend
battery life on battery powered devices which have no requirements
to keep the hub powered during suspend. The driver can also be
configured to leave the hub powered when a wakeup capable USB device
is connected when suspending, and power it off otherwise.

Technically the driver consists of two drivers, the platform driver
described above and a very thin USB driver that subclasses the
generic driver. The purpose of this driver is to provide the platform
driver with the USB devices corresponding to the hub(s) (a hub
controller may provide multiple 'logical' hubs, e.g. one to support
USB 2.0 and another for USB 3.x).

BUG=b:120563698
TEST=on trogdor:
  echo enabled > /sys/devices/platform/soc@0/a6f8800.usb/a600000.dwc3/xhci-hcd.13.auto/power/wakeup
  suspend with USB mouse connected
    => USB hub stays powered on
  suspend without USB mouse connected
    => USB hub is powered off

Change-Id: I7c9a1f1d6ced41dd8310e8a03da666a32364e790
Co-developed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Signed-off-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2104490
Reviewed-by: Douglas Anderson <dianders@chromium.org>