UPSTREAM: drm/i915: Do YCbCr 444->420 conversion via DP protocol converters
For platforms that can't do native 4:2:0 outout we may still be
able to do it by getting the DP->HDMI protocol converter to
perform the 4:4:4->4:2:0 downsamling for us. In this case we
have to configure our hardware to output YCbCr 4:4:4, which we've
already hooked up so all we need to do is flip the switch.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200904115354.25336-19-ville.syrjala@linux.intel.com
Reviewed-by: Lyude Paul <lyude@redhat.com>
Acked-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 181567aa9f0d297804f1ea5d3ff4ba4518e05f2c)

BUG=b:175077993
TEST=Tested on eve-kernelnext

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I6053029f40c036681985b8d36fbf73cde4955861