UPSTREAM: thunderbolt: Retry DROM read once if parsing fails
Kai-Heng reported that sometimes DROM parsing of ASUS PA27AC Thunderbolt 3
monitor fails. This makes the driver to fail to add the device so only
DisplayPort tunneling is functional.

It is not clear what exactly happens but waiting for 100 ms and retrying
the read seems to work this around so we do that here.

Link: https://bugzilla.kernel.org/show_bug.cgi?id=206493
Reported-by: Kai-Heng Feng <kai.heng.feng@canonical.com>
Tested-by: Kai-Heng Feng <kai.heng.feng@canonical.com>
Cc: stable@vger.kernel.org
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry-picked from commit f022ff7bf377ca94367be05de61277934d42ea74)

BUG=b:179702302
TEST=No TBT/USB4 regressions
TEST=No regressions on display over TBT

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: If6b66cce9b8e780769d629062221cd51f144cd04
Signed-off-by: Chintan Patel <chintan.m.patel@intel.corp-partner.google.com>
Signed-off-by: Daniel Campello <campello@chromium.org>