sched/walt: introduce trace_sched_ravg_window_change
Currently we use a printk_deferred to notify when the window size
changes. However, this adds a complication while debugging because it
requires looking at the dmesg output. Instead, we introduce a trace
hook, as this can be easily enabled for collections, and can help
further improve scheduler debugging tools.

Change-Id: I6809589748eef606403940567bd8c66a774b93ea
Signed-off-by: Shaleen Agrawal <shalagra@codeaurora.org>